+++ nodes/knxUltimateHueLight.js
@@ -3,21 +3,6 @@ module.exports = function (RED) {
   const hueColorConverter = require('./utils/hueColorConverter')
 
 
-  async function getLightState(node, _lightID) {
-    return new Promise((resolve, reject) => {
-      try {
-        if (node !== null && node.serverHue !== null && node.serverHue.hueManager !== null) {
-          node.serverHue.hueManager.getLight(_lightID).then(ret => {
-            node.currentHUEDevice = ret[0]
-            resolve(ret)
-          })
-        }
-      } catch (error) {
-        reject(error)
-      }
-    })
-  }
-
   function knxUltimateHueLight(config) {
     RED.nodes.createNode(this, config)
     const node = this
@@ -42,16 +27,7 @@ module.exports = function (RED) {
     node.formatmultiplyvalue = 1
     node.formatnegativevalue = 'leave'
     node.formatdecimalsvalue = 2
-
-
-
-    // Read the state of the light and store it in the holding object
-    try {
-      if (config.hueDevice !== undefined && config.hueDevice !== '') getLightState(node, config.hueDevice)
-    } catch (error) {
-    }
-
-
+    node.currentHUEDevice = {}
 
     // Used to call the status update from the config node.
     node.setNodeStatus = ({ fill, shape, text, payload }) => {
@@ -75,9 +51,9 @@ module.exports = function (RED) {
             if (msg.payload.data > 0) {
               let dimDirection = 'down'
               dimDirection = msg.payload.decr_incr === 1 ? 'up' : 'down'
-              node.startDim(dimDirection)
+              node.startDimStopper(dimDirection)
             } else {
-              node.startDim('stop')
+              node.startDimStopper('stop')
             }
             break
           case config.GALightBrightness:
@@ -91,7 +67,7 @@ module.exports = function (RED) {
             const gamut = node.currentHUEDevice.color.gamut_type || null
             const retXY = hueColorConverter.ColorConverter.rgbToXy(msg.payload.red, msg.payload.green, msg.payload.blue, gamut)
             const bright = hueColorConverter.ColorConverter.getBrightnessFromRGB(msg.payload.red, msg.payload.green, msg.payload.blue)
-            bright > 0 ? state = { on: { on: true }, dimming: { brightness: bright }, color: { xy: retXY } } : state = { on: { on: false } }
+            state = bright > 0 ? { on: { on: true }, dimming: { brightness: bright }, color: { xy: retXY } } : { on: { on: false } }
             node.serverHue.hueManager.setLightState(config.hueDevice, state)
             break
           default:
@@ -105,15 +81,15 @@ module.exports = function (RED) {
     node.timerDim = undefined
     node.dimDirection = {}
     node.timeoutDim = 0
-    node.startDim = function (_direction) {
-      clearInterval(node.timerDim)
+    node.startDimStopper = function (_direction) {
+      if (node.timerDim !== undefined) clearInterval(node.timerDim)
       if (_direction === "stop") return
       switch (_direction) {
         case 'up':
-          node.dimDirection = { dimming_delta: { action: 'up', brightness_delta: 5 } }
+          node.dimDirection = { dimming_delta: { action: 'up', brightness_delta: 10 } }
           break
         case 'down':
-          node.dimDirection = { dimming_delta: { action: 'down', brightness_delta: 5 } }
+          node.dimDirection = { dimming_delta: { action: 'down', brightness_delta: 10 } }
           break
         default:
           break
@@ -122,7 +98,7 @@ module.exports = function (RED) {
         node.timeoutDim += 1
         if (node.timeoutDim > 100) { node.timeoutDim = 0; clearInterval(node.timerDim) }
         node.serverHue.hueManager.setLightState(config.hueDevice, node.dimDirection)
-      }, 50);
+      }, 300);
     }
 
 
@@ -140,7 +116,7 @@ module.exports = function (RED) {
             // ISE Connect Hue emulation, send brightness
             knxMsgPayload.topic = config.GALightBrightnessState
             knxMsgPayload.dpt = config.dptLightBrightnessState
-            _event.on.on === true ? knxMsgPayload.payload = 100 : 0
+            knxMsgPayload.payload = _event.on.on === true ? 100 : 0
             // Send to KNX bus
             if (knxMsgPayload.topic !== '' && knxMsgPayload.topic !== undefined) node.server.writeQueueAdd({ grpaddr: knxMsgPayload.topic, payload: knxMsgPayload.payload, dpt: knxMsgPayload.dpt, outputtype: 'write', nodecallerid: node.id })
           }
@@ -160,7 +136,7 @@ module.exports = function (RED) {
             // ISE Connect Hue emulation, send true/false to switch state
             knxMsgPayload.topic = config.GALightState
             knxMsgPayload.dpt = config.dptLightState
-            _event.dimming.brightness > 0 ? knxMsgPayload.payload = true : false
+            knxMsgPayload.payload = _event.dimming.brightness > 0 ? true : false
             // Send to KNX bus
             if (knxMsgPayload.topic !== '' && knxMsgPayload.topic !== undefined) node.server.writeQueueAdd({ grpaddr: knxMsgPayload.topic, payload: knxMsgPayload.payload, dpt: knxMsgPayload.dpt, outputtype: 'write', nodecallerid: node.id })
 
@@ -180,7 +156,20 @@ module.exports = function (RED) {
     if (node.serverHue) {
       node.serverHue.removeClient(node)
       node.serverHue.addClient(node)
+      setInterval(() => {
+        // Get light state
+        try {
+          if (node !== null && node.serverHue !== null && node.serverHue.hueManager !== null) {
+            node.serverHue.hueManager.getLight(config.hueDevice).then(ret => {
+              node.currentHUEDevice = ret[0]
+            })
+          }
+        } catch (error) {
+          throw (error)
+        }
+      }, 1000);      
     }
+
 
     node.on('input', function (msg) {